requirements.txt

discord.py==2.4.0
python-dotenv==1.0.1
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
openai==1.51.0
